feature(score) add a high Scores display
- Add a ScoreState class to display a list of HiScore,
- Add the Google API registering capability to,
- Add capability to read/write score list from Google Games account

fix #36
