+++ src/main/resources/res/messages.properties
@@ -3,6 +3,7 @@ TitleState.label.start=Start game
 TitleState.label.options=Options
 TitleState.label.quit=Quit
 PlayState.title=Play
+ScoreState.title=Scores
 OptionState.title=Options
 TitleState.label.copyright=GDJ105 - (c) SnapGames 2018
 language=English
